+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Ver 0.4.3 (Next release)
 
   * #48 [When tab title is truncated onhover does not show whole title](https://github.com/georgeOsdDev/react-draggable-tab/issues/48)
+  * #50 [Unable to add styles or css classes to the tab content div](https://github.com/georgeOsdDev/react-draggable-tab/issues/50)
 
 ### Ver 0.4.2